% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_in_window.R
\name{detect_in_window}
\alias{detect_in_window}
\title{Detect Visits before Matching a Fuzzy Sub-Pattern}
\usage{
detect_in_window(
  use_pattern,
  window_width = 4L,
  threshold = 3L,
  match_is = c("+", "-")
)
}
\arguments{
\item{use_pattern}{A character string showing the daily, by visit, or weekly}

\item{window_width}{How wide should the inspection window be? Defaults to
four visits.}

\item{threshold}{How many successes or failures are required within the
specified window? Defaults to three.}

\item{match_is}{What constitutes the outcome of interest? \code{"+"}
represents positive UDS as the visit event of interest, while \code{"-"}
is negative.}
}
\value{
A two column data frame with (a) column \code{time} representing the
number of visits \strong{before} the first window with a match and (b)
column \code{event} indicating if the match occurs in any window of the use
pattern.
}
\description{
Given a use pattern string, use fuzzy logic to detect if a
sub-pattern of interest is present within a specified detection window and
how many visits are required to find said match.
}
\details{
This function can be used to calculate time-to-event (survival)
metrics for the subjects in treatment. For example, the default arguments
represent a definition of relapse ("three or more positive UDS within a
four-week window"). Thus, the output of this function under default values
would be time until first relapse and a relapse indicator.
}
\examples{
  # Find the first relapse event
  detect_in_window("o-o+++")
  
  # Find the first positive UDS
  detect_in_window("o-o+++", window_width = 1L, threshold = 1L)
  
}
