% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_retention.R
\name{measure_retention}
\alias{measure_retention}
\title{Measure Length of Use Pattern before Dropout}
\usage{
measure_retention(use_pattern, missing_is = "o")
}
\arguments{
\item{use_pattern}{A character string showing the daily, by visit, or weekly
substance use pattern for a single subject}

\item{missing_is}{Which single character is used to mark missing UDS in a
use pattern string? Defaults to \code{"o"}.}
}
\value{
An integer measuring the number of visits before the subject was lost
to follow-up. If the subject's entire use pattern is missing, then this
will be 0.
}
\description{
Find the number of visits in the pattern until the last subject
contact
}
\examples{
  pattern_char <- "__++++*o-------+--+-o-o-o+o+oooooo"
  measure_retention(pattern_char)
  
  pattern2_char <- "ooooooooooo"
  measure_retention(pattern2_char)
  
}
