% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_missing.R
\name{recode_missing}
\alias{recode_missing}
\title{Recode Missing or Ambiguous UDS in a Subject Use Pattern}
\usage{
recode_missing(use_pattern, missing_is = "o", missing_as = c("+", "", "-"))
}
\arguments{
\item{use_pattern}{A character string showing the daily, by visit, or weekly
substance use pattern for a single subject}

\item{missing_is}{Which single character is used to mark missing UDS in a
use pattern string? Defaults to \code{"o"}.}

\item{missing_as}{How should missing UDS be treated? Defaults to marking the
subject as positive for that missing period. Options are \code{"+"} for
positive, \code{""} for ignore, or \code{"-"} for negative (note that this
last option should only be used in case where the "missing" UDS is a
regular and expected artefact of the study design and the preceeding and
following UDS are both negative).}
}
\value{
A character string with all missing UDS values (marked as \code{"o"}
unless a different value is supplied to \code{missing_is}) replaces by the
value supplied to \code{missing_as}.
}
\description{
Replace all missing UDS \code{"o"} in a use pattern string
}
\examples{
  pattern_char <- "_____+++++*o-------+--+-o-o-o+o+"
  recode_missing(pattern_char)
  
}
