% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_visits.R
\name{impute_missing_visits}
\alias{impute_missing_visits}
\title{Naively Impute Missing Visits}
\usage{
impute_missing_visits(
  use_pattern,
  method = c("locf", "locfD", "mode"),
  missing_is = "o",
  mixed_is = "*",
  tiebreaker = "+"
)
}
\arguments{
\item{use_pattern}{A character string showing the daily, by visit, or weekly
substance use pattern for a single subject}

\item{method}{Which naive imputation method should be used? Current supported
options are \code{"locf"} (last observation carried forward),
\code{"locfD"} (last observation carried forward until dropout), and
\code{"mode"} (most common non-missing value).}

\item{missing_is}{Which single character is used to mark missing UDS in a
use pattern string? Defaults to \code{"o"}.}

\item{mixed_is}{Which single character is used to mark mixed UDS (both
positive and negative UDS for the visit block) in a use pattern string?
Defaults to \code{"*"}. When imputing by the mode, all mixed result UDS
will be assigned the \code{tiebreaker} value in order to calculate the mode
but will remain unchanged in the returned use pattern string.}

\item{tiebreaker}{In the event of ties between two modes, should positive or
negative UDS be the mode? Defaults to positive (\code{"+"}).}
}
\value{
A use pattern string the same length as \code{use_pattern} with
missing values imputed according to the chosen imputation method.
}
\description{
Given a use pattern string with missing visits, make naive
imputations for each missing visit
}
\details{
If you would like to replace all UDS for missing visits with a
single, pre-specified value (such as positive), please use
\code{\link{recode_missing_visits}} instead. Furthermore, there will most
likely still be missing values in the use pattern even after imputation.
This would occur if all the values are missing, if the first values of the
use pattern are missing (if LOCF is used), or if the first and/or last
values of the use pattern are missing (if LOCF-D) is used. Because of this,
you may need to call \code{\link{recode_missing_visits}} in a pipeline
after this function to replace or remove the remaining non-imputable
missing visits.
}
\examples{
  pattern_char <- "__++++*o-------+--+-o-o-o+o+oooooo"
  impute_missing_visits(pattern_char)
  impute_missing_visits(pattern_char, method = "locfD")
  impute_missing_visits(pattern_char, method = "mode")
  
  pattern2_char <- "ooooooooooo"
  impute_missing_visits(pattern2_char)
  
}
